Refactor NetworkManager
NetworkManager creates new PacketTransceiver each query. This way even if we run multiple queries in parallel, we won't have any problem.
That we decoupled the implementation of PacketTransceiver from NetworkManager.
NetworkManager does not know how PacketTransceiverFactory is providing PacketTransceiver, if it's the same instance or different.
It's only job is to ask for a PacketTransceiver for each query and notify the listener.
It's job is unaffected even if ExecutorService changes to sequential to parallel.