revisit root servers response
see scratchpad for more details.

It took me time to express what i want to do.
But at the end I came up with some good code.
There are no nested if's in getNextQuery in DNSResolver and I love that.
Sometimes it takes time and it's worth giving.
I know in future it may change but for now it's good enough.